refactored models back to single files, handled imports - tested & working still
