+++ app/models/tag.py
@@ -1,0 +1,17 @@
+from .db import db
+from .tables import product_tag
+
+
+class Tag(db.Model):
+    __tablename__ = "tags"
+    id = db.Column(db.Integer, primary_key=True)
+    tag = db.Column(db.String(30), nullable=False)
+
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "tag": self.tag,
+        }
+
+    tag_product = db.relationship(
+        "Product", secondary=product_tag, back_populates='product_tag')
